alive is as a single internal component
